Keep MCP search tests under the file-size ratchet.
Move the unscoped ranking coverage out of search.node.test.ts and stop
using the mcp:<name> domain as an identity field so a package tagged
only "mcp" does not wrap every connected server.
