+++ packages/worker/src/mcp/tools/search-mcp-servers.node.test.ts
@@ -5,8 +5,11 @@ import { formatEntityDetailMarkdown, parseEntityRef } from './search-format.ts'
 import {
 	buildMcpServerToolIndex,
 	findSynthesizedMcpServer,
+	findWrappingPackageForMcpServer,
 	listSynthesizedMcpServers,
 } from './search-mcp-servers.ts'
+import { searchUnified } from './search.ts'
+import { type PackageSearchRow } from './search-types.ts'
 
 function createHomeRegistry() {
 	const homeServer = {
@@ -84,6 +87,29 @@ test('MCP server entities resolve by name or domain and list tools with instruct
 	})
 	const tools = buildMcpServerToolIndex(server!)
 	expect(tools.map((tool) => tool.toolName)).toEqual(['set_pin', 'list_lights'])
+	expect(server!.identityFields).toEqual(['home', 'home'])
+	expect(
+		findWrappingPackageForMcpServer(server!, [
+			{
+				record: {
+					kodyId: 'notes',
+					name: '@user/notes',
+					tags: ['mcp'],
+				},
+			} as PackageSearchRow,
+		]),
+	).toBeNull()
+	expect(
+		findWrappingPackageForMcpServer(server!, [
+			{
+				record: {
+					kodyId: 'home',
+					name: '@user/home',
+					tags: ['home'],
+				},
+			} as PackageSearchRow,
+		]),
+	).toMatchObject({ kodyId: 'home' })
 
 	const detail = formatEntityDetailMarkdown({
 		type: 'mcp-server',
@@ -112,4 +138,110 @@ test('MCP server entities resolve by name or domain and list tools with instruct
 		capabilityCount: 2,
 		instructions: 'Use set_pin after unlocking the island router.',
 	})
+})
+
+test('unscoped search ranks the MCP server instead of dumping every remote tool', async () => {
+	const homeServer = {
+		serverId: 'server-home',
+		serverName: 'home',
+		kodyName: 'home',
+	}
+	const toolNames = [
+		'set_pin',
+		'get_pin',
+		'list_lights',
+		'set_light',
+		'get_thermostat',
+		'set_thermostat',
+		'lock_door',
+		'unlock_door',
+		'screenshot',
+		'run_script',
+		...Array.from({ length: 20 }, (_, index) => `home_tool_${String(index)}`),
+	]
+	const registry = buildCapabilityRegistry([
+		{
+			name: 'mcp:home',
+			description:
+				'Control lights, locks, and the island router PIN on the home LAN.',
+			keywords: ['mcp', 'integration'],
+			capabilities: toolNames.map((toolName) => ({
+				name: `mcp:home:${toolName}`,
+				domain: 'mcp:home',
+				description:
+					toolName === 'set_pin'
+						? 'Set the island router PIN.'
+						: `Home automation tool ${toolName}.`,
+				keywords: ['home', toolName.replaceAll('_', ' ')],
+				readOnly: false,
+				idempotent: false,
+				destructive: false,
+				source: 'mcp-server' as const,
+				mcpServer: {
+					...homeServer,
+					mcpToolName: toolName,
+					toolName,
+				},
+				inputSchema: { type: 'object' as const, properties: {} },
+				inputTypeDefinition: 'type HomeToolInput = Record<string, never>',
+				handler: async () => null,
+			})),
+		},
+	])
+	const optionalRows = {
+		packageRows: [],
+		userSecretRows: [],
+		userValueRows: [],
+		userIntegrationRows: [],
+	}
+
+	const unscopedHome = await searchUnified({
+		env: {} as Env,
+		query: 'home',
+		limit: 15,
+		registry,
+		optionalRows,
+	})
+	expect(
+		unscopedHome.matches.filter((match) => match.type === 'capability'),
+	).toEqual([])
+	expect(
+		unscopedHome.matches.filter((match) => match.type === 'mcp-server'),
+	).toEqual([
+		expect.objectContaining({
+			type: 'mcp-server',
+			kodyName: 'home',
+			domain: 'mcp:home',
+			capabilityCount: toolNames.length,
+			instructions:
+				'Control lights, locks, and the island router PIN on the home LAN.',
+		}),
+	])
+	expect(unscopedHome.guidance).toContain('home:mcp-server')
+
+	const unscopedPin = await searchUnified({
+		env: {} as Env,
+		query: 'set pin',
+		limit: 15,
+		registry,
+		optionalRows,
+	})
+	expect(
+		unscopedPin.matches.filter((match) => match.type === 'capability'),
+	).toEqual([])
+	expect(unscopedPin.matches.some((match) => match.type === 'mcp-server')).toBe(
+		true,
+	)
+
+	const domainListing = await searchUnified({
+		env: {} as Env,
+		query: '',
+		limit: 50,
+		domain: 'mcp:home',
+		registry,
+		optionalRows,
+	})
+	expect(
+		domainListing.matches.filter((match) => match.type === 'capability'),
+	).toHaveLength(toolNames.length)
 })
