+++ packages/worker/src/mcp/tools/search.node.test.ts
@@ -1275,110 +1275,6 @@ test('searchUnified inlines call shapes for the top three capability matches onl
 	expect(listTop).not.toHaveProperty('inputTypeDefinitionTruncated')
 })
 
-test('unscoped search ranks the MCP server instead of dumping every remote tool', async () => {
-	const homeServer = {
-		serverId: 'server-home',
-		serverName: 'home',
-		kodyName: 'home',
-	}
-	const toolNames = [
-		'set_pin',
-		'get_pin',
-		'list_lights',
-		'set_light',
-		'get_thermostat',
-		'set_thermostat',
-		'lock_door',
-		'unlock_door',
-		'screenshot',
-		'run_script',
-		...Array.from({ length: 20 }, (_, index) => `home_tool_${String(index)}`),
-	]
-	const registry = buildCapabilityRegistry([
-		{
-			name: 'mcp:home',
-			description:
-				'Control lights, locks, and the island router PIN on the home LAN.',
-			keywords: ['mcp', 'integration'],
-			capabilities: toolNames.map((toolName) => ({
-				name: `mcp:home:${toolName}`,
-				domain: 'mcp:home',
-				description:
-					toolName === 'set_pin'
-						? 'Set the island router PIN.'
-						: `Home automation tool ${toolName}.`,
-				keywords: ['home', toolName.replaceAll('_', ' ')],
-				readOnly: false,
-				idempotent: false,
-				destructive: false,
-				source: 'mcp-server' as const,
-				mcpServer: {
-					...homeServer,
-					mcpToolName: toolName,
-					toolName,
-				},
-				inputSchema: { type: 'object' as const, properties: {} },
-				inputTypeDefinition: 'type HomeToolInput = Record<string, never>',
-				handler: async () => null,
-			})),
-		},
-	])
-
-	const unscopedHome = await searchUnified({
-		env: {} as Env,
-		query: 'home',
-		limit: 15,
-		registry,
-		optionalRows: emptyOptionalSearchRows,
-	})
-	const unscopedHomeCapabilities = unscopedHome.matches.filter(
-		(match) => match.type === 'capability',
-	)
-	expect(unscopedHomeCapabilities).toEqual([])
-	expect(
-		unscopedHome.matches.filter((match) => match.type === 'mcp-server'),
-	).toEqual([
-		expect.objectContaining({
-			type: 'mcp-server',
-			kodyName: 'home',
-			domain: 'mcp:home',
-			capabilityCount: toolNames.length,
-			instructions:
-				'Control lights, locks, and the island router PIN on the home LAN.',
-		}),
-	])
-	expect(unscopedHome.guidance).toContain('home:mcp-server')
-
-	const unscopedPin = await searchUnified({
-		env: {} as Env,
-		query: 'set pin',
-		limit: 15,
-		registry,
-		optionalRows: emptyOptionalSearchRows,
-	})
-	expect(
-		unscopedPin.matches.filter((match) => match.type === 'capability'),
-	).toEqual([])
-	expect(unscopedPin.matches.some((match) => match.type === 'mcp-server')).toBe(
-		true,
-	)
-
-	const domainListing = await searchUnified({
-		env: {} as Env,
-		query: '',
-		limit: 50,
-		domain: 'mcp:home',
-		registry,
-		optionalRows: emptyOptionalSearchRows,
-	})
-	expect(
-		domainListing.matches.filter((match) => match.type === 'capability'),
-	).toHaveLength(toolNames.length)
-	expect(domainListing.matches.map((match) => match.type)).toEqual(
-		Array.from({ length: toolNames.length }, () => 'capability'),
-	)
-})
-
 test('settleWithBudget uses an absolute launch deadline and degrades safely', async () => {
 	await expect(
 		settleWithBudget(Promise.resolve('ready'), 25),
